fix: ActivityDto
- Se soluciono el bug que no mostraba los puntos de reputacion en las actividades

+++ src/main/java/ar/edu/unq/desapp/grupoD022021/backenddesappapi/dto/ActivityDto.java
@@ -10,7 +10,7 @@ public class ActivityDto  {
         int cantidad;
         String fullNameUser;
         int numberOperations;
-        double reputation;
+        int reputation;
         String action;
         String emailUser;
 
@@ -87,7 +87,7 @@ public class ActivityDto  {
         return reputation;
     }
 
-    public void setReputation(double reputation) {
+    public void setReputation(int reputation) {
         this.reputation = reputation;
     }
 }
